Rework/Comment a few elements of the Polls feature
