+++ sdk/src/main/kotlin/io/rover/sdk/ui/blocks/poll/TextPollViewModel.kt
@@ -1,19 +1,12 @@
 package io.rover.sdk.ui.blocks.poll
 
-import android.graphics.Paint
-import io.rover.sdk.data.domain.Color
-import io.rover.sdk.data.domain.Font
-import io.rover.sdk.data.domain.TextAlignment
 import io.rover.sdk.data.domain.TextPollBlock
-import io.rover.sdk.platform.getFontAppearance
-import io.rover.sdk.platform.mapToFont
+import io.rover.sdk.data.getFontAppearance
 import io.rover.sdk.services.MeasurementService
 import io.rover.sdk.streams.PublishSubject
 import io.rover.sdk.ui.RectF
-import io.rover.sdk.ui.asAndroidColor
 import io.rover.sdk.ui.blocks.concerns.background.BackgroundViewModelInterface
 import io.rover.sdk.ui.blocks.concerns.layout.Measurable
-import io.rover.sdk.ui.blocks.concerns.text.FontAppearance
 import io.rover.sdk.ui.concerns.BindableViewModel
 
 internal class TextPollViewModel(
